+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Required configuration fields are now handled by Baker rather than by each component. [#41](https://github.com/AdRoll/baker/pull/41)
 - filter: add TimestampRange filter [#46](https://github.com/AdRoll/baker/pull/46)
 - filter: add ReplaceFields filter [#49](https://github.com/AdRoll/baker/pull/49)
+- filter: add Timestamp filter [#54](https://github.com/AdRoll/baker/pull/54)
 
 ### Changed